+++ controller/userCtrl/user_controller.go
@@ -1,4 +1,4 @@
-package controller
+package userCtrl
 
 import "github.com/labstack/echo/v4"
