+++ controller/userCtrl/user_controller_impl.go
@@ -1,4 +1,4 @@
-package controller
+package userCtrl
 
 import (
 	"github.com/labstack/echo/v4"
